Clippy (#5)
* add teme2gcrf function

* doc updates

* use clippy for rust check, and remove all warnings

* fix python docs to match "as_" instead of "to_"

* up versions in Cargo.toml

* mv Pipfile and requirements.txt to python basedir

+++ src/pybindings/pykepler.rs
@@ -202,7 +202,7 @@ impl PyKepler {
     }
 
     fn __getstate__(&mut self, py: Python) -> PyResult<PyObject> {
-        let mut state = [0 as u8; 48];
+        let mut state = [0; 48];
         state[0..8].clone_from_slice(&self.inner.a.to_le_bytes());
         state[8..16].clone_from_slice(&self.inner.eccen.to_le_bytes());
         state[16..24].clone_from_slice(&self.inner.incl.to_le_bytes());
